// frontend/src/types/ServerMessage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SessionState } from "./SessionState";

/**
 * Messages sent from server to client
 */
export type ServerMessage = { "type": "TestMessage", text: string, } | { "type": "StateSync", session: SessionState, } | { "type": "VideoMetadata", width: number | null, height: number | null, duration_seconds: number | null, };