// frontend/src/types/ClientMessage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PlaybackSpeedUpdate } from "./PlaybackSpeedUpdate";
import type { PlaybackUpdate } from "./PlaybackUpdate";
import type { SessionState } from "./SessionState";
import type { VolumeUpdate } from "./VolumeUpdate";

/**
 * Messages sent from client to server
 */
export type ClientMessage = { "type": "UpdatePlaybackPosition" } & PlaybackUpdate | { "type": "UpdatePlaybackSpeed" } & PlaybackSpeedUpdate | { "type": "UpdateVolume" } & VolumeUpdate | { "type": "SyncState" } & SessionState;